refactor(agent): centralize classic battle event dispatch
