+++ CONTRIBUTING.md
@@ -155,6 +155,15 @@ When contributing code, especially in gameplay or input-related areas, follow th
 
 ---
 
+## 🔄 Migration Notes: Classic Battle Dispatch
+
+- All Classic Battle events now dispatch through a single pathway: `eventDispatcher.dispatchBattleEvent`.
+- Modules should import from `eventDispatcher.js` instead of `orchestrator.js` or the removed `battleDispatcher.js`.
+- The orchestrator registers the machine via `eventDispatcher.setMachine`; call this once during initialization.
+- Tests can mock `eventDispatcher.dispatchBattleEvent` to observe transitions.
+
+---
+
 ## 📎 Related AI Agent Docs
 
 - `AGENTS.md` – Tasks, prompts, and expectations for agents
