+++ tests/helpers/classicBattle/interruptHandlers.test.js
@@ -6,7 +6,7 @@ let modalElement;
 vi.mock("../../../src/helpers/battleEngineFacade.js", () => ({
   battleEngine: { interruptMatch: vi.fn() }
 }));
-vi.mock("../../../src/helpers/classicBattle/orchestrator.js", () => ({
+vi.mock("../../../src/helpers/classicBattle/eventDispatcher.js", () => ({
   dispatchBattleEvent: vi.fn()
 }));
 vi.mock("../../../src/helpers/setupScoreboard.js", () => ({
@@ -81,7 +81,7 @@ describe("initInterruptHandlers", () => {
     );
     const { battleEngine } = await import("../../../src/helpers/battleEngineFacade.js");
     const { dispatchBattleEvent } = await import(
-      "../../../src/helpers/classicBattle/orchestrator.js"
+      "../../../src/helpers/classicBattle/eventDispatcher.js"
     );
     const { showMessage, clearTimer } = await import("../../../src/helpers/setupScoreboard.js");
     const { stop: stopScheduler, cancel } = await import("../../../src/utils/scheduler.js");
@@ -113,7 +113,7 @@ describe("initInterruptHandlers", () => {
     );
     const { battleEngine } = await import("../../../src/helpers/battleEngineFacade.js");
     const { dispatchBattleEvent } = await import(
-      "../../../src/helpers/classicBattle/orchestrator.js"
+      "../../../src/helpers/classicBattle/eventDispatcher.js"
     );
     const { showMessage } = await import("../../../src/helpers/setupScoreboard.js");
 
@@ -136,7 +136,7 @@ describe("initInterruptHandlers", () => {
     );
     const { battleEngine } = await import("../../../src/helpers/battleEngineFacade.js");
     const { dispatchBattleEvent } = await import(
-      "../../../src/helpers/classicBattle/orchestrator.js"
+      "../../../src/helpers/classicBattle/eventDispatcher.js"
     );
     const { showMessage } = await import("../../../src/helpers/setupScoreboard.js");
 
@@ -162,7 +162,7 @@ describe("initInterruptHandlers", () => {
     );
     const { battleEngine } = await import("../../../src/helpers/battleEngineFacade.js");
     const { dispatchBattleEvent } = await import(
-      "../../../src/helpers/classicBattle/orchestrator.js"
+      "../../../src/helpers/classicBattle/eventDispatcher.js"
     );
     const { showMessage } = await import("../../../src/helpers/setupScoreboard.js");
