+++ tests/helpers/classicBattle/scheduleNextRound.fallback.test.js
@@ -29,7 +29,7 @@ describe("scheduleNextRound fallback timer", () => {
     vi.doMock("../../../src/helpers/classicBattle/skipHandler.js", () => ({
       setSkipHandler: vi.fn()
     }));
-    vi.doMock("../../../src/helpers/classicBattle/battleDispatcher.js", () => ({
+    vi.doMock("../../../src/helpers/classicBattle/eventDispatcher.js", () => ({
       dispatchBattleEvent: vi.fn()
     }));
     vi.doMock("../../../src/helpers/classicBattle/battleEvents.js", () => ({
