+++ tests/helpers/classicBattlePage.syncScoreDisplay.test.js
@@ -7,7 +7,9 @@ beforeEach(() => {
   vi.resetModules();
   vi.doUnmock("../../src/helpers/settingsStorage.js");
   vi.doMock("../../src/helpers/classicBattle/orchestrator.js", () => ({
-    initClassicBattleOrchestrator: vi.fn(),
+    initClassicBattleOrchestrator: vi.fn()
+  }));
+  vi.doMock("../../src/helpers/classicBattle/eventDispatcher.js", () => ({
     dispatchBattleEvent: vi.fn().mockResolvedValue()
   }));
 });
