+++ tests/helpers/timerService.ordering.test.js
@@ -36,7 +36,7 @@ describe("timerService timeout ordering", () => {
       if (eventName === "timeout") return timeoutPromise;
       return Promise.resolve();
     });
-    vi.doMock("../../src/helpers/classicBattle/battleDispatcher.js", () => ({
+    vi.doMock("../../src/helpers/classicBattle/eventDispatcher.js", () => ({
       dispatchBattleEvent: dispatchSpy
     }));
